refactor: 更新转码相关文本为“转码”，优化代码结构

+++ lib/models/app_configs.dart
@@ -200,9 +200,7 @@ class AppConfigs {
   }
 
   static int _clampThemeHue(int hue) {
-    return hue
-        .clamp(AppConstants.themeHueMin, AppConstants.themeHueMax)
-        .toInt();
+    return hue.clamp(AppConstants.themeHueMin, AppConstants.themeHueMax);
   }
 
   static int? _legacySeedColorToHue(Object? raw) {
